refactor: clean up expense history auto-update implementation
- Remove duplicate MonthYearSelector functions, create shared component
- Implement automatic expense list updates using Flow observation
- Replace hardcoded values with ValidationConstants.LATEST_EXPENSES_COUNT
- Fix deprecated Date methods with modern Calendar approach
- Remove unused functions and lifecycle management code
- Add proper month/year formatting for Latest Expenses title
- Improve code maintainability and consistency
