+++ app/src/main/java/com/example/budget/ui/budget/BudgetScreen.kt
@@ -2,6 +2,7 @@ package com.example.budget.ui.budget
 
 import androidx.compose.foundation.layout.*
 import com.example.budget.ui.components.BeautifulSelector
+import com.example.budget.ui.components.MonthYearSelector
 import com.example.budget.ui.components.ConfirmationMessage
 import com.example.budget.data.ValidationConstants
 import androidx.compose.foundation.lazy.LazyColumn
@@ -424,46 +425,5 @@ fun BudgetScreen(
     }
 }
 
-@Composable
-private fun MonthYearSelector(
-    selectedMonth: Int,
-    selectedYear: Int,
-    onMonthChange: (Int) -> Unit,
-    onYearChange: (Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val months = DateConstants.MONTHS
-    val years = DateConstants.getAvailableYears()
-
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        // Month Selector
-        Box(modifier = Modifier.weight(1f)) {
-            BeautifulSelector(
-                label = "Month",
-                value = months.find { it.second == selectedMonth }?.first ?: "January",
-                options = months.map { it.first },
-                onSelectionChange = { selectedMonthName ->
-                    val monthValue = months.find { it.first == selectedMonthName }?.second ?: 1
-                    onMonthChange(monthValue)
-                }
-            )
-        }
-
-        // Year Selector
-        Box(modifier = Modifier.weight(1f)) {
-            BeautifulSelector(
-                label = "Year",
-                value = selectedYear.toString(),
-                options = years.map { it.toString() },
-                onSelectionChange = { selectedYearString ->
-                    onYearChange(selectedYearString.toInt())
-                }
-            )
-        }
-    }
-}
