+++ app/src/main/java/com/example/budget/ui/components/MonthYearSelector.kt
@@ -1,0 +1,49 @@
+package com.example.budget.ui.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.budget.data.DateConstants
+
+@Composable
+fun MonthYearSelector(
+    selectedMonth: Int,
+    selectedYear: Int,
+    onMonthChange: (Int) -> Unit,
+    onYearChange: (Int) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val months = DateConstants.MONTHS
+    val years = DateConstants.getAvailableYears()
+
+    Row(
+        modifier = modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        // Month Selector
+        Box(modifier = Modifier.weight(1f)) {
+            BeautifulSelector(
+                label = "Month",
+                value = months.find { it.second == selectedMonth }?.first ?: "January",
+                options = months.map { it.first },
+                onSelectionChange = { selectedMonthName ->
+                    val monthValue = months.find { it.first == selectedMonthName }?.second ?: 1
+                    onMonthChange(monthValue)
+                }
+            )
+        }
+
+        // Year Selector
+        Box(modifier = Modifier.weight(1f)) {
+            BeautifulSelector(
+                label = "Year",
+                value = selectedYear.toString(),
+                options = years.map { it.toString() },
+                onSelectionChange = { selectedYearString ->
+                    onYearChange(selectedYearString.toInt())
+                }
+            )
+        }
+    }
+}
