+++ app/src/main/java/com/example/budget/ui/expensehistory/ExpenseHistoryScreen.kt
@@ -19,7 +19,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.budget.data.db.Expense
 import com.example.budget.ui.AppViewModelProvider
 import com.example.budget.ui.components.ConfirmationMessage
-import com.example.budget.ui.components.BeautifulSelector
+import com.example.budget.ui.components.MonthYearSelector
 import com.example.budget.ui.utils.formatCurrency
 import com.example.budget.ui.utils.getDynamicTextStyle
 import com.example.budget.data.DateConstants
@@ -220,44 +220,3 @@ fun ExpenseHistoryItem(
     }
 }
 
-@Composable
-fun MonthYearSelector(
-    selectedMonth: Int,
-    selectedYear: Int,
-    onMonthChange: (Int) -> Unit,
-    onYearChange: (Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val months = DateConstants.MONTHS
-    val years = DateConstants.getAvailableYears()
-
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        // Month Selector
-        Box(modifier = Modifier.weight(1f)) {
-            BeautifulSelector(
-                label = "Month",
-                value = months.find { it.second == selectedMonth }?.first ?: "January",
-                options = months.map { it.first },
-                onSelectionChange = { selectedMonthName ->
-                    val monthValue = months.find { it.first == selectedMonthName }?.second ?: 1
-                    onMonthChange(monthValue)
-                }
-            )
-        }
-
-        // Year Selector
-        Box(modifier = Modifier.weight(1f)) {
-            BeautifulSelector(
-                label = "Year",
-                value = selectedYear.toString(),
-                options = years.map { it.toString() },
-                onSelectionChange = { selectedYearString ->
-                    onYearChange(selectedYearString.toInt())
-                }
-            )
-        }
-    }
-}
